backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.80.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
firebase-admin>=6.5.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
requests>=2.31.0
